Added a smallWindow to display on Settings.json error.
Was testing the ability to add new keys and ran into a noob error where I was deleting the 2nd line but keeping the line above "," at the end of the line. This tells the json to expect another line thus when it does not find the next line it fails at reading the whole file.

Stumped me for a few minutes before I realized it, thus thought it deserved a smallWindow if the settings.ts ever fails to read the settings.json file correctly.

So this now displays a smallWindow on launching the program whenever settings.json fails to load and gives the error it receives.