"background" setting -> "videoBackground" setting.
Realized that it is very likely that both images and video will be supported (ideally, and although would be taxing use at the same time).
So background has been split into videoBackground and imageBackground.

For now only Video background path works, and works exactly as it was working when it was just background path.

Also adjusted the video background path setting fileDialog to only allow selecting video type files. (I think this can be expanded to more filetypes like .mkv etc. but will get around to testing this later).