Very primitive 3 state Background Solution.
3 options for backgrounds to appear:
1. Video background path is valid and video displays.
2. Video not working or not valid -> Show image background if that is valid or displays.
3. Shows the fallback solid color background.

The solution to the image background working without needing to deploy another file protocol is to just save them in AppData/Roaming/AltDesktop/background/ filePath. Thus the filePath is relative to that folder (set it to background.png for a file in that folder called background.png).

This also means that the backgrounds must be copied there, but as this is only for the image backgrounds, the folder should still remain of decent size (not like the videos which can be expected to be gigabytes in size and would quickly expand to become massive).

Not my favorite solution, but this conversion will essentially become automated when you add the file in image background path in a few updates (like how the icon copies and converts).