SaveBackgroundImage on saving settings.
Copies the image at path to AppData/Roaming/AltDesktop/background/ folder, then sets the image background path to the localized path. This primitively automates the saving and path conversion such that when selecting a file or when saving the settings it will apply the image background.

Known issue: SafeFileProtocol by default returns an unknown.png image, which means as long as image background path is not empty, it will show the unknown.png as a background if the background path is incorrect/not an image etc.

Still not 100% satisfied with this implementation. Might look into an alternative way of loading non-safeFileProtocol images for backgrounds alone. As if you swap backgrounds a lot, even just back and forth this would not catch that unless you were selecting from the /background/ folder thus would duplicate another background.png every time you swap it.