Stop copying existing backgrounds. (images).
Added some special cases for copying which will attempt to find the file in the directory first (even if it has an incremented name "name.png", "name(1).png", "name(2).png" etc.) and return that file name instead of copying the image every time.

note: obviously same file with different names are not checked as this could very likely result in perfomance issues if say someone copied over thousands of images. But very very unlikely to cause problems with this short check when assuming a vast majority of files do not share a name.