Footers no longer block fields (EditIcon and settings).
Adjusted the css to show the scroll bar as soon as the footer would start to cover up one component.
Previously it would only show up when the fields would stretch below the global bottom of the window (when it was already hiding behind the footer). Thus a shorter window would never show the last field.

Also fixed some text which should not have been selectable.