Icon font size in Settings.tsx
Added a font size field in Settings.tsx window which allows you to change the default font size.

Also adjusted the fontSize to technically allow being undefined and updated it such that if it is undefined it defaults to 16.

This allows setting the Icon font size to "" which just means use default (16).