WIP background preview.
ipcHandler now sends the update to the background.tsx where it attempts to temporarily override the background to update to the changed values.

Known issues: Closing without saving does not reload the background (thus can get stuck with a previewed background).
Can't preview both video/image backgrounds at the same time. (Reason for this is each field only sends an update for their field, and if it's not passed as an update it defaults back to the saved settings for the other background field).

Both seem fixable, just need some work/testing.