Improved fallback to image if video fails to load.
Threw a literal brick wall (400,000kb/s bitrate 85gb h265 video at the video background, and found that the videoBackground fails... (lmao).

On the related note I found that it was not properly falling back to a correct ImageBackground when this happens due to calling getBackgroundImagePath twice (whoops). Fixed that unified the image fallback into a single function, instead for uniformity.

Also another reason it was not properly falling back to Image was it was getting stuck in the retries due to low level errors. So at least for now the retries has been removed, especially as it seems any failure to load the video is immediately obvious, and unlikely that retries would actually do anything.