Selecting ImageBackground and dropping now previews
Fixed the problem with previewing for image backgrounds.
Problem was it would only save to /backgrounds/ when it was saved, thus could not preview as it would attempt previewing an unobtainable file. (due to safe-file protocol).

This does mean that it now saves any images you select/drop into the settings in /backgrounds/.

Known issue manual copy/paste or typing in the field does not preview correctly. Might just loosen the strictness of the safe-file protocol. Most of the protocol strictness has not been because of actual security concerns but more so for practice and understanding as this is my first electron based large project.