Fixed both Background Paths empty preview.
Fixed a case where both Background Paths empty in settings would cause either imageBackground or videoBackground to display instead of the base fallback.

Problem with this was predominantly that if a path was empty it was not sending both paths. Thus the fetchBackground/fallbackToImage assumed that if one was not passed it should use the settings.json value preview.