Added extractFileIcon ipc call.
Added extract-file-icon package.

Attached it to a new ipc call extractFileIcon which gets the icon of the passed path, and saves it (for now) to AppDataBasePath (AppData/Roaming/AltDesktop).

From basic testing works for getting the direct icon, works with .lnks, and works for getting the default program icons if the file itself does not have an icon.