Alternative approach to getting file icon.
Ran into some major problems with extractFileIcon, where it would work great, but only work part of the time. For instance when starting up the program it worked ~30% of the time, the other ~70% of the time it would save a completely blank .png to the directory, never logged errors or gave any hints as to why it was failing.

This is now a custom attempt at getting the file icon.