Combination attempt.
Creating a workflow for an attempt to use the extract-file-icon package, and then fallback to the python executable on failure. This uses the faster extract-file-icon which also gets better default program icons, but if it fails, fallsback to the (currently only supports exe files exe_to_image executable (exe_to_image.py).