file_to_image updated for Default Program icons.
Updated file_to_image to accept a type "exe" or "default" (for now).

Added default_icon_to_image() which now gets the default program's icon, saves it an returns its path.

Tested electron-builder exe on a non-dev environment pc and still seems to be working correctly. Just need to implement a bunch of other special cases like a way to get favicons, and workarounds for applications which use .lnks which link to a folder which then uses a app.ico instead of embedding the icon into the executable..... (discord)...

Also more fallbacks etc.