extractFileIcon now returns string[] of saved files.
Since it will be possible to generate multiple icons, it is best to adjust this to just pass back an array of strings for each possible icon. Then we will implement some small menu which will preview the different icons, and allow the user to select which auto generated one to use.