BrowserIcon fallback. (if weblink does not return favicon).
Added a fallback when extractFileIcon is called with a webLink, which will get the default browser path and an icon from it and return it if the google api fails to load or returns nothing.

Example use for this would be no internet connection, would always return this, or a bad path i.e. "google" instead of "google.com"