Ico no longer direct save to data folder.
Best to set up the icons to not directly copy unless the user actually sets them. This is harder to do with icons fetched via specific methods, but is very easy for .ico files since they have a defined source. Whereas, something like getting the image from an exe, default program, favicon etc. all basically require at minimum saving to a file, so it makes sense just to save it to the data directory.