SelectIconWindow and smallWindow CSS fixes.
Added a SelectIconWindow which appears after clicking "Generate Icon" button in EditIcon.

It opens a SelectIconWindow which shares a lot of CSS and design with SmallWindow.tsx.

It loads safeImages of all the returned image paths from generateIcon ipc call, then allows the user to select one. As of now just logs "Selected icon: (path to icon)".

Updated CSS for smallWindow to ensure that the header always appears.
The button responses for smallWindow also now always appear.
Adjusted smallWindow spawned size to be taller.
Adjusted the content (small window message) to be scrollable if the message is too large to fit.

Fixed "save-button" for EditIcon/Settings to only have margin-left when "generate-button" also exists (keeping it centered in Settings menu).

Known issue: SelectIconWindow (if not all smallWindows alt-tab to a non-AltDesktop window upon closing) I had fixed this for subWindows, so need to fix this for these smallWindows/SelectIconWindow.