Fix smallWindow closing focusing non AltDesktop Window.
Fixed a problem where closing a smallWindow or SelectIconWindow would cause previous focus target to be lost, thus would default back to focusing last (non altDesktop) window upon closing.

The problem is basically with this, is the last focused item gets confused when closing, so all it needs is it's last focus target to be the mainWindow, then when it closes it has a valid last focused target. It's only when it has an invalid process/windowID for the last focus target that it defaults to first alt+tab window.