SmallWindow devTools toggle.
Added a smallWindow tools toggle in the devtools menu (in header when launched as dev).

Simple checkbox (defaults unchecked). When checked will open a detached devtools menu when opening a smallWindow or SelectIconWindow. Very useful for adjusting CSS and understanding the CSS in development. Ideally any actual problems are logged.