Corrected saving to attempt to save all "image" filetypes for icon.image.
Fixed a problem where a file like .ico was not included in allowedExtensions, thus would not attempt SaveIconImage ipc call.

This now calls as long as the filetype returns with "image/" prefix.

Still some other problems with .icos in general and some other type issues that need to be worked out.