Fixed preview to work with icons that do not yet exist.
Fixed previewing for icon.images to create a tempIcon and place it into the iconsMap to allow a brand new icon to preview what the icon image would look like.

Updated reloadIcon ipc call to always send a reload-icon request.

Updated DesktopGrid when receiving the message to remove the icon from the map if the icon update received is null (special case to delete the icon). (only sends null if reloadIcon ipc is called with no instance of the icon at (row,col) existing in the desktopIcons.json.)