Fixed special characters interfering with baseName detection.
Fixed special characters in filePaths like () {} etc. in filepaths when ipc saveIconImage or saveBackgroundImage. This previously lead to selecting the same original file, saving another copy with a different name in the /data/ path.

Example case file name: app - Copy (2).ico (Assuming this same file is selected and 1 file with the same name exists in the data folder).
Previously this would create a new "app - Copy (2)(1).ico" file.
Due to the special characters not detecting the existing file with the same base name.

Now:
Properly detects same base name, compares actual file, if same file then just returns the path of the saved image/background. If not saves it and passes back that saved path.