Implemented defaultIconSize when setting icon sizes.
DesktopGrid and SafeImage now correctly load the defaultIconSize and adjust all icons to match it (if they do not have explicitly set width/height.

Obvious problems which will be fixed:
Icons can outgrow their containers (will probably scale the size of icon boxes with the size of the icons).
Not yet a way to change it in the UI only in the settings.json
Requires reload in order for the change to display.