iconBox size scales based on defaultIconSize.
Updated the icon box to scale based on the size of the icons set in settings.json.

i.e. previously setting the defaultIconSize to 128 would cause them to drastically overflow their assigned box spilling over into other icons etc.

This now adjusts the box the icon resides in to also scale based on the size of the icon, making it smaller or larger and ensuring that the icon stays within its own box (unless its offset or size has been manually changed).