Saving settings now guarantees defaultSettings.
Updated saveSettingsData to call ensureDefaultSettings() after saving. This ensures that when the user saves with an empty input (a sign to revert to default value) it properly saves with the defaultSetting value for that key.

Example: Saves Default icon size setting with "" (left blank to revert to default"

Next time user opens settings it displays default font size: 64

Basically just uses the blank as a temporary indication user wants default value, then when the user edits it next, it preloads with the actual default value that they can then edit right away instead of having to guess the default value and adjust from there.