Fixed icon sizes not resizing on settings changed.
Fixed the issue where SafeImage was getting defaultIconSize instead of just passing it via DesktopGrid.

This led to icons not resizing based on the settings defaultIconSize as it fetched it once on launch then would require a restart/Refresh Desktop to load an updated defaultIconSize. Its now updated in desktopGrid which passes this updated value to SafeImage.