PreviewGridUpdate background logic.
Added the sendPreviewGridUpdate from settings.tsx when updating a field.

Added the ipc previewGridUpdate which receives that message and sends a message to the DesktopGrid.

Added receiver for the DesktopGrid which gets this update and the updated values.

Did not implement the actual previewing, just the entire system for sending the preview values from the Settings.tsx -> main program -> back to DesktopGrid.tsx