Closing without saving reloads Grid to remove previewed changes.
Closing Settings without saving now sends a message to reload the DesktopGrid when there have been changes made to the
defaultIconSize or defaultFontSize. To revert back to the saved settings by reverting the previewed changes.

Also fixed the webContents.sed for reload-background which was not changed in the last commit to match the updated name scheme.