Only show color picker on clicking the preview.
Can now select Font Color and manually type/manually set it blank (to revert to default when this is implemented).

Did a lot of experimenting trying to get more advanced color pickers, but the only electron supported ones I could find very outdated and struggle to work. There are plenty of JS/node based ones but none of them seem to work without major change to accomodate them.

Will just put this on the backburner and come back to trying to improve the color selector when time allows. Realized I should prioritize general usage over a small UI problem.

Only real issue with the chromium default is that the eye dropper doesn't show cursor in mainWindow, and does not keep the cursor special while outside the EditIcon window (other programs/desktop etc.) and right click effectively works the same as a left click, when intuitively I as a user would assume that a right click means cancel which is not the case.