Previews defaultFontColor
Now previews the defaultFontColor when changing it in Settings menu.

Need to cut back on the logging/updating however as this is doing a lot of updating due to passing back updates as complete (always passes all update args, and calls fetchFontSize, fetchIconSize, and fetchFontColor regardless of which has been changed).