Only sends the grid update for changed setting.
Reduced the update for grid settings to only send back the 1 update it changes. This works nicely as it is basically just changing a local variable to an override which stays like that till reloaded.

Might export this setup to Background.tsx as it could benefit from this.

All updates should only be called with one update type and be fine sending back just that update. (should never have a case where it sends back updates["defaultIconSize", "defaultFontSize"] etc. it should just have 1 at a time.