Fix SFP pathname and SafeImage requesting "none" or " "
Fixed a problem with safeFileProtocol checking only hostname instead of pathname for "unknown".
This would lead to getting the unknown.png request return via fileNotExist instead of the getUnknownImageResponse(). (not a massive deal, everything still worked but this produced a logger warning).

Fixed SafeImage special cases "none" or " " from sending requests for the actual files "none" or " " (They returned empty anyways, but this stops the request in general).