Swapping between two header types.
Can now swap between the two header types. (Requires refresh or relaunch for now).

Only difference at the moment is that the header gets smaller if you set it to borderless. (more will be changed soon). This is just a test/confirm case to ensure that swapping the CSS/class name is viable.

Also pondering on splitting up the App.css to simple component style files. To make things easier to find, edit, and avoid global definition problems like the header CSS that was affecting all <header> objects without me realizing it (was changed yesterday to .main-header, and again in this commit to window type class headers.