1st Attempt Borderless window.
Consolidated the two header files into Header.css
Decreased size of Borderless Window Header bar to minimum size to fit all components.
In Borderless the header hides by default.
Added a header-hotspot object in the top right which causes the header to show when hovered.
(Basically move your mouse to the top right of the window and the header will appear.)
Header also appears when hovering the hover-opacity-item.
Resized hover-opacity-item to match the height of the header (but fully across the top of the window.)

Window Type "Windowed" no longer has access to the hover-opacity-item (it is hidden behind the header).

Basic thought process for this:
1. Windowed: works as expected program, user can easily assume how the window works and it always shows.
2. Borderless: more advanced, header is hidden for a more immersive desktop, hovering over the top of the program sets background to be transparent (to see through the program (but not the icons (might add a setting for this)), normal header still appears in top right (only when mouse nearby).