Borderless while maximized removes resizing options.
Removed the resizing along the borders of the program when program is Maximized and set to Borderless.

The big reason for this is so that you can just move your mouse up all the way to the top of the monitor, and it will still trigger the hover-opacity-item. If it remains resizable, then moving the mouse all the way to the top of the monitor instead readies resizing the window, and does not trigger the hover-opacity-item. Thus you would need to get it close to the edge but not on the edge of the monitor to trigger it.