unMaximized window always uses windowed mode.
Tried just implementing a small drag box for the borderless window type, but ran into the bug of all onHover() functionality breaking with transparent: true browserWindows + webkit-app-region: drag;

So instead the Borderless option will be reserved for Maximized only, and when unmaximized it will use the normal windowed header.
This easily allows for unmaximizing, then dragging the header to reposition it to another screen, or just using it while unmaximized.