header-hotspot resizes based on header size (partially).
Work in progress of making the header-hotspot resize based on the width of the header.

I think this can be done more efficiently/better cover bases through a different way, so going to work on that, and if it doesn't work this is a start to a solution as well.