Simplicity is King (Removed Borderless hover-hotspot).
Tried many times to just create a a simple/efficient hover-hotspot which would satisfy the following:
1. Shows header while hovered.
2. Allows right click-throughs to desktop-grid.
3. Allows left click-throughs to desktop-grid.

Really hard to satisfy all 3. In the last commit and for a while I just worked on creating a hover-hotspot which satisfied #1 and scaled with the size of the header when in borderless.
Got a few working attempts with that, fixing a few bugs created in the built executable version, over a few re-designs.

Then the problem came of requirement #2 and #3. Tried many different iterations by creating a HoverHotspot component, trying to embed the div in DesktopGrid component, and a few other implementations. None could fully satisfy all 3 requirements.

The closest I have gotten is the DesktopGrid controlling a <div> for header-hotspot, and using ipc to "force-show" the header. This satisfied #1 and #2, but failed on left click pass-through. Thus any icons in the top right would not be launchable. This also added quite the overhead if you just spam moved your mouse over the header-hotspot (in/out of it). As it used an ipc and a webContents.send to force show the header. And I would have to redo the size constraints as this was resizing to match the header, even when the header was "windowed" thus would span the entire top row of icons. So it would require even more work.

With all these negatives it simply makes the most sense to just keep it simple, this is only 1 Setting which is not turned on by default, introducing all these bugs and possible hitching due to accidental spam hover/unhover, it simply makes sense to simplify it.

Current thought process is that the "Borderless" setting is for advanced users, and either I can add some quick tutorial or a warning/explanation somewhere and the advanced user should understand that you just have to hover the top right (or the opacity-item) to activate the header.

Also in my attempts to send clicks to the DesktopGrid I realized that DesktopGrid did not have a className for it's component, and thus added "desktop-grid" class