transparent window arg, now defined by windowType on launch.
For now the windowType controls if the browserWindow is transparent or not.

For instance launching with windowType = "WINDOWED" will not allow transparency.
Conversely launching with windowType = "BORDERLESS" allows transparency.

Two things this affects.
1. non-transparent windows headers can be double clicked to maximize/unmaximize.
2. transparent windows can use the hover-opacity-item to make the program transparent partially or fully.

In the future the transparency will be probably a hidden setting, where if the user has windowType = "borderless" or has an added hover-opacity-item it will try to launch as transparent.
And switching to borderless or adding a transparency item will advise the user to restart to allow transparency.

This better maintains the nice qualities of a windowed mode for the non-advanced users who expect it, and allows for transparency for the advanced users who want it.