showIcons = false no longer shows icon contextMenu.
Fixed a problem where it would show the contextMenu for the desktop, or the contextMenu for the icon when showIcons was set to false.

This now shows its own "hideIcons" contextMenu.type where the first option is Show Icons.

This prevents the user from filling there screens then suddenly they can't show icons because every icon shows its contextMenu. And this prevents weird possible issues such as creating a new icon while icons are hidden.