Fix ImageBg saveBgImage() on already relative paths.
Fixed a case where saving Settings with an imageBackground path that is relative, was still attempting to find and save the image.

Now for relative paths (already saved backgrounds), it skips attempting to save it as it should 1. already exist in the relative paths or 2. not a valid file.

Drag/drop, copy/pasting, or using the magnifying glass, the filepath that is generated for imageBackground will always be an absolute path, the only time you get relativePath is manually typing in the filename.