Massive simplification of Background, (No image/video, now just Background).
Extreme simplification of Background.tsx.

Removed all logic for imageBackground/videoBackground.
Now only uses 1 "Background Path" to set the background.
This background path can be image or video, it solves for both.

This is a move towards eventually removing this manually set path. In the future there will be a dedicated Window for selecting background. Which will allow you to select from between any you have used before. On selecting one it will keep/store the path to the file which will then send to the Background.tsx thus the user won't have to deal with the setting field path at all, they will just select a saved background and it will load for them.

Note: some debugging for videos was removed, like ffprobe checking the video on error. Will add this back soon.