Added back debugging and analytics for Video Backgrounds.
Added back in a lot of the videoMetrics, warnings and logging for detecting problems with a video background.

Also improved on the detection of other codecs besides hvc1.
Added the message for hev1, and added a generic window on error which logs the codec name and codec tag.

Also fixed a log statement which was of type Error which commonly occured when the user swapped from a video background to another background. This now is of type info and gives a hint that it can be ignored if occurred after swapping from a video background (if no other errors occur).