Fix long word/strings overflowing smallWindow.
Fixed an issue where a really long unbroken filepath could overflow, creating a horizontal scroll bar.
Further the problem with the horizontal scroll bar was that if it appeared it was still cutting off some of the left side of the word/string.

Thus it has just been replaced with a stricter word wrapping to avoid this situation entirely.