Added Background button to header.
Added a button "background-icon" to header which will provide a way to open the "background" window (not yet implemented).

for now just logs "background button clicked".

Also removed the desktopIcons_example as this is a bit outdated, and the program has completely managed this for some time, you haven't needed to create your own for a while.