Pre-emptive rename to EditBackground.
Pre-emptively renaming the last 2 commits references to "Background" to instead reference it as "Edit Background".

Realized that referencing it as Background might get confusing since there is already a "Background.tsx" which exists, so further references to this editing background window will reference it fully as "editBackground", "edit-background", "Edit Background" etc.