Loading data from backgrounds.json into Edit Backgrounds.
Added an ipc call getBackgroundSummaries which returns some basic information from the bg.json for the indexed background.

Right now returns and displays:
Name/ID
Description
Full filePath (to background (should be in same folder)).
iconPath which gets displayed as a SafeImage (loads and displays the actual image).
Tags: [] Array of tags.

Still need to adjust a lot of stuff, and create some data structures for sorting, but this is a good test just to ensure that we can get the latest data and deliver it to the UI