Condensed grid and now scales to fit.
Condensed the grid to just 5px, greatly increasing how they stack next to other background-grid-items.

Increased Name area to 2 lines max.
background-grid-item scales to fit width of total grid. (For example the grid-items scale larger if there is empty space which is not enough to fit the next grid-item.)
Background-icons now scale based on the width/height of the background-grid-item.