Fixed safeImage always adding /data/[row,col] when called with row, col.
Fixed an issue where calling safeImage with a row, col value would always add the /data/[$row},${col}] to the path, even if it was already called with an absolute file path.

This particularly broke on generating default program icons for non-executables/images.
Also updated the generateIcon to return the fileName for the generated icon instead of the full path to avoid a situation like this entirely, so as far as a I know there are no cases where safeImage gets called both with a row and col AND an absolutePath. But if the case does exist it should still work as it no longer adds the data/[row,col] to the beginning of the request.