Selecting background from EditBackground now previews.
If you have a valid backgrounds/ folder which matches the style of the provided PLACE_IN_BACKGROUNDS folder, now when clicking on to select the background it calls to preview the background.

Does not actually change background and does not reload background after closing EditBackground.
Also this only works as long as the "filename" is within the same folder as the bg.json.
Looking into a few ways to not strictly require this. Possibly the masterPaths I have hinted at which would allow essentially different install locations similar to "backgrounds/" folder this draws from. Other than that, might make it work with shortcuts to reference fully external video files. (Say for instance you have a large file you don't want multiple copies of in an external file, instead of being forced to copy it to AltDesktops folders, maybe it can get it via shortcut).