Setting Background Path now only takes indexed File IDs.
Instead of saving full file background paths to "Background Path" setting it now just saves the ID of the index owner file.

For example if you put the PLACE_IN_BACKGROUNDS 000001 file in your AltDesktop/backgrounds/ ("000001" file goes here).
Upon launch it will index this and show in your backgrounds.json file.

Then you can set the Background Path to 000001 and it will auto fetch the background and set it correctly.

Also updated the Edit Backgrounds window to send updates via bg.id instead of the full file path.

Still need to implement saving on closing Edit Backgrounds to actually save the file to the "background path" setting. Also will remove the Background Path setting (from the UI only) as it will be fully controlled by the program and the user won't need to set it manually. Also need to implement a reload on closing Edit Backgrounds to ensure that whatever was saved is up to date.