Added support for shortcut video/image backgrounds.
Can now place a shortcut file in /backgrounds/filename_here/
folders and use it as such.

This could help with allowing a user to save just a shortcut instead of needing to save the raw file itself inside this path.
When saving one and creating the bg.json (in program when it will support it), will need to add a check to save just a shortcut for video backgrounds, image backgrounds can probably just default copy it.

Also fixed some old ipcMain.ons to use ipcMainHandle instead. (old logging ipc calls).