Fix icon contextMenus calling normal desktop contextMenus.
Fixed a case where right clicking a desktopIcon would trigger both the "icon" contextMenu and the "desktop" contextMenu, thus always showing the desktop contextMenu.
This caused right clicks on a DesktopIcon which had an offset outside of its home to trigger the contextmenu of where the click was, not the icons home.