isSubWindowActive -> getSubWindowTitle ipc call.
Adjusted the isSubWindowActive to instead return the subWindowTitle. As such changed the name to "getSubWindowTitle". This returns "" if no subWindow found which equates to boolean false in cases like:
if (await window.electron.getSubWindowTitle()) {}
So it works the same as it used to, but now returns the title of the subWindow which is much more useful.