getSubWindowTitle now re-centers on current screen.
Fixed getSubWindowTitle re-centering to re-center on the current screen, instead of the main screen/screen that the window was originally opened on.

For instance if you opened the subWindow, then moved the main window, and tried to right click, or open another subWindow it would re-center the subWindow on the screen where it was opened, not the current screen.