Basic EditBackground window.
Added a window EditBackground.tsx which is opened on dropping a file onto Background Select.

This window works similar to subWindow for now, but is not a singleton and can be opened along side other subWindows.

All it currently does is use the settings CSS and display Drop FIle: (filePath).

Set up to also work with no filePath (can be called via await window.electron.openEditBackground()  ).

In the future this will essentially convert an external file like image/video to a file within the /backgrounds/ directory so it can be indexed. This will also create the bg.json and an icon preview image. This will also handle saving image based backgrounds to the same file, and most likely shortcuts to the video or copying the video to the directory (with a toggle for which).

Also updated the getFileType() to return "directory" for folders.
Looking into possibly supporting dropping files in the future, but need a lot more brainstorming about what is expected/how it works.

Also as I was writing this I am thinking I might revert EditBackground to a normal singleton SubWindow. Depends on how I go about previewing the background/icon and how Background Select gets updated.