Update SubWindow and EditBackground is now SubWindow.
Updated the overall singleton strategy of openSubWindow to asynchronously create the windows so that it can properly close the activeSubWindow when called and re-open a new subWindow. (Previously had a race condition which prevented this from actually re-opening a subWindow).

strengthened closeActiveSubWindow to callback (required for the async createWindow()) when the activeSubWindow is closed and force close if needed.

EditBackground is now a normal subWindow.
openEditBackgroundWIndow is now removed as it is not used.
Moved the showDevtools for subWindows into the subWindow itself (simplified and was having trouble opening it for EditBackgroundWindow).