EditBackground now previews background.
EditBackground window now previews the File Path as a background.

Updated the update-background-preview message handler to allow for previewing absolute file paths instead of just IDs. Still converts from ID to FilePath if not called with an absolute file path.

On closing the editBackground, now reloads the background. (to undo the preview).