Fixed swapping to/from (" ", "none") to normal image paths.
Fixed swapping an icon path from an icon/unknown.svg to a blank icon when changing the file path to " " or "none".

Also fixed when swapping from a blank icon (saved as " " or "none") to a normal image/unknown.svg.

Both previously required desktop reloads to properly show.
This bug was introduced a few days ago when safeImage was simplified, but this was an oversight where the isEmptyImage was only ever set once on launch and never updated. It now updates in the useEffect on imagePath change.