Failure in bg.json still loads in BackgroundSelect.
If it encounters an error trying to read the bg.json it now still loads it into BackgroundSelect window, just with a default no name, image, description, tags.

This will allow essentially showing the user that its detected, but not working. (might add some css to make it more noticeable besides just the unknown.svg).

Also when edit background is fully implemented, this will allow the user to basically rebuild their bg.json.