bg.json creation via ipc saveBgJson.
Added an ipc call saveBgJson which saves a BackgroundSummary object to the bg.json at ID's filepath.
Example call:
await window.electron.saveBgJson({
  id: "000007",
  name: "My Backgrounds",
  filename: "background.jpg",
  description: "A background",
  iconPath: "icon.png",
  tags: ["summer"],
  localTags: ["test-tag", "summer-2025"]
});

Saves the file:
{
  "public": {
    "name": "My Backgrounds",
    "filename": "background.jpg",
    "description": "A background",
    "tags": [
      "summer"
    ],
    "icon": "icon.png"
  },
  "local": {
    "tags": [
      "test-tag",
      "summer-2025"
    ]
  }
}

Still need to attach it to the save button, and do a bunch of testing. But as of now this will create the folder even if it doesn't exist either.