EditBackground now manages a BackgroundSummary.
Saving now somewhat works. But needs a lot of work.

Some things that need to be fixed.
Right now just calls with id = "xyz" on fileDrop in BackgroundSelect to open EditBackground.
Doesn't save file locally, a shortcut/copying the image to the background/id folder.
And several other things need to be fixed.