Now saves files to /backgrounds/id/ folder.
Now when saving Edit Background it saves the files (as long as filename is a video or an image, and as long as icon is an image (shows warnings otherwise))

Note: for now it is saving the entire video to the /backgrounds/id/ folder.
In the future I plan to add a checkbox for saving these to local or creating a shortcut, with the default for image to save the actual file, and default for video to save a shortcut.

This will mean that in the future if you move the video file that you use as your background it will break the shortcut. But doing it like this allows you to keep videos externally without needing to create a copy inside the backgrounds/id/ folder itself.